V1.9 memcopy
